HIJACKING

+++ gframe/menu_handler.cpp
@@ -26,7 +26,6 @@
 #include <IGUITabControl.h>
 #include <IGUITable.h>
 #include <IGUIWindow.h>
-#include "address.h"
 
 namespace ygo {
 
@@ -39,8 +38,8 @@ static void UpdateDeck() {
 	const auto totsize = deck.main.size() + deck.extra.size() + deck.side.size();
 	if(totsize > max_deck_size)
 		return;
-	BufferIO::Write<uint32_t>(pdeck, static_cast<uint32_t>(deck.main.size() + deck.extra.size()));
-	BufferIO::Write<uint32_t>(pdeck, static_cast<uint32_t>(deck.side.size()));
+	BufferIO::Write<uint32_t>(pdeck, deck.main.size() + deck.extra.size());
+	BufferIO::Write<uint32_t>(pdeck, deck.side.size());
 	for(const auto& pcard : deck.main)
 		BufferIO::Write<uint32_t>(pdeck, pcard->code);
 	for(const auto& pcard : deck.extra)
@@ -108,12 +107,12 @@ bool MenuHandler::OnEvent(const irr::SEvent& event) {
 	case irr::EET_GUI_EVENT: {
 		irr::gui::IGUIElement* caller = event.GUIEvent.Caller;
 		int id = caller->getID();
-		if(mainGame->wRules->isVisible() && (id != BUTTON_RULE_OK && id != CHECKBOX_EXTRA_RULE && id != COMBOBOX_DUEL_RULE && id != EDITBOX_TEAM_COUNT))
+		if(mainGame->wRules->isVisible() && (id != BUTTON_RULE_OK && id != CHECKBOX_EXTRA_RULE && id != COMBOBOX_DUEL_RULE))
 			break;
 		if(mainGame->wMessage->isVisible() && id != BUTTON_MSG_OK &&
 		   prev_operation != ACTION_UPDATE_PROMPT
 		   && prev_operation != ACTION_SHOW_CHANGELOG
-#if EDOPRO_LINUX && (IRRLICHT_VERSION_MAJOR==1 && IRRLICHT_VERSION_MINOR==9)
+#if defined(__linux__) && !defined(__ANDROID__) && (IRRLICHT_VERSION_MAJOR==1 && IRRLICHT_VERSION_MINOR==9)
 		   && prev_operation != ACTION_TRY_WAYLAND
 #endif
 		   )
@@ -195,11 +194,11 @@ bool MenuHandler::OnEvent(const irr::SEvent& event) {
 			}
 			case BUTTON_JOIN_HOST: {
 				try {
-					const auto parsed = epro::Host::resolve(mainGame->ebJoinHost->getText(), mainGame->ebJoinPort->getText());
+					auto parsed = DuelClient::ResolveServer(mainGame->ebJoinHost->getText(), mainGame->ebJoinPort->getText());
 					gGameConfig->lasthost = mainGame->ebJoinHost->getText();
 					gGameConfig->lastport = mainGame->ebJoinPort->getText();
 					mainGame->dInfo.secret.pass = mainGame->ebJoinPass->getText();
-					if(DuelClient::StartClient(parsed.address, parsed.port, 0, false)) {
+					if(DuelClient::StartClient(parsed.first, parsed.second, 0, false)) {
 						mainGame->btnCreateHost->setEnabled(false);
 						mainGame->btnJoinHost->setEnabled(false);
 						mainGame->btnJoinCancel->setEnabled(false);
@@ -268,17 +267,16 @@ bool MenuHandler::OnEvent(const irr::SEvent& event) {
 					}
 					gGameConfig->gamename = mainGame->ebServerName->getText();
 					gGameConfig->serverport = mainGame->ebHostPort->getText();
-					mainGame->gBot.Refresh(gGameConfig->filterBot * (mainGame->cbDuelRule->getSelected() + 1), gGameConfig->lastBot);
 					if(!NetServer::StartServer(host_port))
 						break;
-					const auto ip = 0x100007F; //127.0.0.1 in network byte order
-					if(!DuelClient::StartClient({ &ip, epro::Address::INET }, host_port)) {
+					if(!DuelClient::StartClient(0x100007F /*127.0.0.1 in network byte order*/, host_port)) {
 						NetServer::StopServer();
 						break;
 					}
 					DuelClient::is_local_host = true;
 					mainGame->btnHostConfirm->setEnabled(false);
 					mainGame->btnHostCancel->setEnabled(false);
+					mainGame->gBot.Refresh(gGameConfig->filterBot * (mainGame->cbDuelRule->getSelected() + 1), gGameConfig->lastBot);
 				}
 				break;
 			}
@@ -309,10 +307,14 @@ bool MenuHandler::OnEvent(const irr::SEvent& event) {
 				break;
 			}
 			case BUTTON_HP_KICK: {
+				int id = 0;
+				while(id < 6) {
+					if(mainGame->btnHostPrepKick[id] == caller)
+						break;
+					id++;
+				}
 				CTOS_Kick csk;
-				csk.pos = 0;
-				while (csk.pos < 6 && mainGame->btnHostPrepKick[csk.pos] != caller)
-					csk.pos++;
+				csk.pos = id;
 				DuelClient::SendPacketToServer(CTOS_HS_KICK, csk);
 				break;
 			}
@@ -444,19 +446,7 @@ bool MenuHandler::OnEvent(const irr::SEvent& event) {
 					if(mainGame->gBot.LaunchSelected(port, mainGame->dInfo.secret.pass))
 						break;
 				} catch(...) {}
-				mainGame->PopupMessage(gDataManager->GetSysString(12122).data());
-				break;
-			}
-			case BUTTON_BOT_COPY_COMMAND: {
-				try {
-					int port = std::stoi(gGameConfig->serverport);
-					const auto params = mainGame->gBot.GetParameters(port, mainGame->dInfo.secret.pass);
-					if(params.size()) {
-						Utils::OSOperator->copyToClipboard(mainGame->gBot.GetParameters(port, mainGame->dInfo.secret.pass).data());
-						mainGame->stACMessage->setText(gDataManager->GetSysString(12121).data());
-						mainGame->PopupElement(mainGame->wACMessage, 20);
-					}
-				} catch(...) {}
+				mainGame->PopupMessage(L"Failed to launch windbot");
 				break;
 			}
 			case BUTTON_EXPORT_DECK: {
@@ -561,42 +551,29 @@ bool MenuHandler::OnEvent(const irr::SEvent& event) {
 			}
 			case BUTTON_YES: {
 				mainGame->HideElement(mainGame->wQuery);
-				switch(prev_operation) {
-#if EDOPRO_LINUX && (IRRLICHT_VERSION_MAJOR==1 && IRRLICHT_VERSION_MINOR==9)
-				case ACTION_TRY_WAYLAND: {
+#if defined(__linux__) && !defined(__ANDROID__) && (IRRLICHT_VERSION_MAJOR==1 && IRRLICHT_VERSION_MINOR==9)
+				if(prev_operation == ACTION_TRY_WAYLAND) {
 					gGameConfig->useWayland = 1;
 					mainGame->SaveConfig();
 					Utils::Reboot();
 				}
 #endif
-				case BUTTON_DELETE_REPLAY: {
+				if(prev_operation == BUTTON_DELETE_REPLAY) {
 					if(Replay::DeleteReplay(Utils::ToPathString(mainGame->lstReplayList->getListItem(prev_sel, true)))) {
 						mainGame->stReplayInfo->setText(L"");
 						mainGame->lstReplayList->refreshList();
 					}
-					break;
-				}
-				case BUTTON_DELETE_SINGLEPLAY: {
+				} else if(prev_operation == BUTTON_DELETE_SINGLEPLAY) {
 					if(Utils::FileDelete(Utils::ToPathString(mainGame->lstSinglePlayList->getListItem(prev_sel, true)))) {
 						mainGame->stSinglePlayInfo->setText(L"");
 						mainGame->lstSinglePlayList->refreshList();
 					}
-					break;
-				}
-				case ACTION_UPDATE_PROMPT: {
+				} else if(prev_operation == ACTION_UPDATE_PROMPT) {
 					gClientUpdater->StartUpdate(Game::UpdateDownloadBar, mainGame);
 					mainGame->HideElement(mainGame->wMainMenu);
 					mainGame->PopupElement(mainGame->updateWindow);
-					break;
-				}
-				case ACTION_SHOW_CHANGELOG: {
-					Utils::SystemOpen(EPRO_TEXT("https://github.com/edo9300/edopro/releases"), Utils::OPEN_URL);
-					break;
-				}
-				case ACTION_ACKNOWLEDGE_HOST: {
-					DuelClient::JoinFromDiscord();
-					break;
-				}
+				} else if (prev_operation == ACTION_SHOW_CHANGELOG) {
+					Utils::SystemOpen(EPRO_TEXT("https://github.com/edo9300/edopro/releases"));
 				}
 				prev_operation = 0;
 				prev_sel = -1;
@@ -604,7 +581,7 @@ bool MenuHandler::OnEvent(const irr::SEvent& event) {
 			}
 			case BUTTON_NO: {
 				switch(prev_operation) {
-#if EDOPRO_LINUX && (IRRLICHT_VERSION_MAJOR==1 && IRRLICHT_VERSION_MINOR==9)
+#if defined(__linux__) && !defined(__ANDROID__) && (IRRLICHT_VERSION_MAJOR==1 && IRRLICHT_VERSION_MINOR==9)
 				case ACTION_TRY_WAYLAND:
 					gGameConfig->useWayland = 0;
 					mainGame->SaveConfig();
@@ -663,9 +640,10 @@ bool MenuHandler::OnEvent(const irr::SEvent& event) {
 				int sel = mainGame->lstHostList->getSelected();
 				if(sel == -1)
 					break;
-				const auto& selection = DuelClient::hosts[sel];
-				mainGame->ebJoinHost->setText(fmt::to_wstring(selection.address).data());
-				mainGame->ebJoinPort->setText(fmt::to_wstring(selection.port).data());
+				int addr = DuelClient::hosts[sel].ipaddr;
+				int port = DuelClient::hosts[sel].port;
+				mainGame->ebJoinHost->setText(epro::format(L"{}.{}.{}.{}", addr & 0xff, (addr >> 8) & 0xff, (addr >> 16) & 0xff, (addr >> 24) & 0xff).data());
+				mainGame->ebJoinPort->setText(fmt::to_wstring(port).data());
 				break;
 			}
 			case LISTBOX_REPLAY_LIST: {
@@ -697,7 +675,7 @@ bool MenuHandler::OnEvent(const irr::SEvent& event) {
 				mainGame->btnShareReplay->setEnabled(true);
 				std::wstring repinfo;
 				time_t curtime = replay.pheader.base.timestamp;
-				repinfo.append(epro::format(L"{:%Y/%m/%d %H:%M:%S}\n", fmt::localtime(curtime)));
+				repinfo.append(epro::format(L"{:%Y/%m/%d %H:%M:%S}\n", *std::localtime(&curtime)));
 				const auto& names = replay.GetPlayerNames();
 				for(int i = 0; i < replay.GetPlayersCount(0); i++) {
 					repinfo.append(names[i] + L"\n");
@@ -749,11 +727,11 @@ bool MenuHandler::OnEvent(const irr::SEvent& event) {
 				if(sel == -1)
 					break;
 				try {
-					const auto parsed = epro::Host::resolve(mainGame->ebJoinHost->getText(), mainGame->ebJoinPort->getText());
+					auto parsed = DuelClient::ResolveServer(mainGame->ebJoinHost->getText(), mainGame->ebJoinPort->getText());
 					gGameConfig->lasthost = mainGame->ebJoinHost->getText();
 					gGameConfig->lastport = mainGame->ebJoinPort->getText();
 					mainGame->dInfo.secret.pass = mainGame->ebJoinPass->getText();
-					if(DuelClient::StartClient(parsed.address, parsed.port, 0, false)) {
+					if(DuelClient::StartClient(parsed.first, parsed.second, 0, false)) {
 						mainGame->btnCreateHost->setEnabled(false);
 						mainGame->btnJoinHost->setEnabled(false);
 						mainGame->btnJoinCancel->setEnabled(false);
@@ -855,11 +833,11 @@ bool MenuHandler::OnEvent(const irr::SEvent& event) {
 		case irr::gui::EGET_EDITBOX_CHANGED: {
 			switch(id) {
 			case EDITBOX_PORT_BOX: {
-				const wchar_t* text = caller->getText();
+				std::wstring text = caller->getText();
 				wchar_t filtered[20];
 				int j = 0;
 				bool changed = false;
-				for(int i = 0; text[i] && j < 19; i++) {
+				for(int i = 0; text[i]; i++) {
 					if(text[i] >= L'0' && text[i]<= L'9') {
 						filtered[j] = text[i];
 						j++;
@@ -867,13 +845,12 @@ bool MenuHandler::OnEvent(const irr::SEvent& event) {
 					}
 				}
 				filtered[j] = 0;
-				text = filtered;
-				if(BufferIO::GetVal(text) > 65535) {
-					text = L"65535";
+				if(BufferIO::GetVal(filtered) > 65535) {
+					wcscpy(filtered, L"65535");
 					changed = true;
 				}
 				if(changed)
-					caller->setText(text);
+					caller->setText(filtered);
 				break;
 			}
 			case EDITBOX_TEAM_COUNT: {
@@ -936,7 +913,7 @@ bool MenuHandler::OnEvent(const irr::SEvent& event) {
 					}
 	#undef CHECK
 					mainGame->duel_param |= tcg;
-					for (auto i = 0u; i < sizeofarr(mainGame->chkCustomRules); ++i) {
+					for (int i = 0; i < sizeofarr(mainGame->chkCustomRules); ++i) {
 						bool set = false;
 						if(i == 19)
 							set = mainGame->duel_param & DUEL_USE_TRAPS_IN_NEW_CHAIN;
@@ -953,7 +930,7 @@ bool MenuHandler::OnEvent(const irr::SEvent& event) {
 							mainGame->chkCustomRules[4]->setEnabled(set);
 					}
 					static constexpr uint32_t limits[]{ TYPE_FUSION, TYPE_SYNCHRO, TYPE_XYZ, TYPE_PENDULUM, TYPE_LINK };
-					for (auto i = 0u; i < sizeofarr(mainGame->chkTypeLimit); ++i)
+					for (int i = 0; i < sizeofarr(mainGame->chkTypeLimit); ++i)
 							mainGame->chkTypeLimit[i]->setChecked(mainGame->forbiddentypes & limits[i]);
 				}
 				break;
@@ -1014,7 +991,7 @@ bool MenuHandler::OnEvent(const irr::SEvent& event) {
 				mainGame->UpdateExtraRules();
 				}
 #undef CHECK
-				for(auto i = 0u; i < sizeofarr(mainGame->chkCustomRules); ++i) {
+				for(int i = 0; i < sizeofarr(mainGame->chkCustomRules); ++i) {
 					bool set = false;
 					if(i == 19)
 						set = mainGame->duel_param & DUEL_USE_TRAPS_IN_NEW_CHAIN;
@@ -1031,7 +1008,7 @@ bool MenuHandler::OnEvent(const irr::SEvent& event) {
 						mainGame->chkCustomRules[4]->setEnabled(set);
 				}
 				static constexpr uint32_t limits[]{ TYPE_FUSION, TYPE_SYNCHRO, TYPE_XYZ, TYPE_PENDULUM, TYPE_LINK };
-				for(auto i = 0u; i < sizeofarr(mainGame->chkTypeLimit); ++i)
+				for(int i = 0; i < sizeofarr(mainGame->chkTypeLimit); ++i)
 					mainGame->chkTypeLimit[i]->setChecked(mainGame->forbiddentypes & limits[i]);
 				break;
 			}
@@ -1095,7 +1072,7 @@ bool MenuHandler::OnEvent(const irr::SEvent& event) {
 		}
 		break;
 	}
-#if !EDOPRO_ANDROID && !EDOPRO_IOS
+#ifndef __ANDROID__
 	case irr::EET_DROP_EVENT: {
 		static std::wstring to_open_file;
 		switch(event.DropEvent.DropType) {
